add simple_navbar bootstrap mode
