+++ Gemfile
@@ -4,6 +4,4 @@ gem 'activesupport', '3.2.12'
 gem 'nokogiri'
 gem 'rails'
 gem 'haml'
-gem 'simple-page-compoents',
-    :git => 'git://github.com/mindpin/simple-page-compoents'
 gem 'coveralls', require: false
