+++ lib/simple_navbar/config/current_context.rb
@@ -1,0 +1,10 @@
+require 'singleton'
+
+module SimpleNavbar
+  module Config
+    class CurrentContext
+      attr_accessor :rules, :nav
+      include Singleton
+    end
+  end
+end
