+++ lib/simple_navbar/config/error.rb
@@ -1,0 +1,6 @@
+module SimpleNavbar
+  module Config
+    class UnconfigurationError < Exception;end
+    class UndefinedRuleError < Exception;end
+  end
+end
