--- lib/simple_navbar/current_context.rb
@@ -1,9 +1,0 @@
-# -*- encoding : utf-8 -*-
-require 'singleton'
-
-module SimpleNavbar
-  class CurrentContext
-    attr_accessor :rules, :nav
-    include Singleton
-  end
-end
