--- lib/simple_navbar/error.rb
@@ -1,5 +1,0 @@
-# -*- encoding : utf-8 -*-
-module SimpleNavbar
-  class UnconfigurationError < Exception;end
-  class UndefinedRuleError < Exception;end
-end
