+++ lib/simple_navbar/helpers.rb
@@ -1,0 +1,4 @@
+require 'simple_navbar/helpers/simple_navbar_helper'
+require 'simple_navbar/helpers/simple_breadcrumbs_helper'
+require 'simple_navbar/helpers/simple_navtabs_helper'
+require 'simple_navbar/helpers/quick_filter_bar_helper'
