--- lib/simple_navbar/quick_filter_bar_helpers.rb
@@ -1,13 +1,0 @@
-module SimpleNavbar
-  module QuickFilterBarHelpers
-    # = quick_filter_bar do |builder|
-    #   - builder.group :state, :text => "状态" do |group|
-    #     - group.add "answed",   :text => "已回答"
-    #     - group.add "unanswed", :text => "未回答"
-    def quick_filter_bar(&block)
-      builder = SimpleNavbar::QuickFilterBar::Builder.new(self)
-      block.call(builder)
-      builder.render
-    end
-  end
-end
