+++ lib/simple_navbar/render.rb
@@ -1,0 +1,3 @@
+require 'simple_navbar/render/simple_navbar'
+require 'simple_navbar/render/simple_navtabs'
+require 'simple_navbar/render/quick_filter_bar'
