+++ lib/simple_navbar/render/quick_filter_bar.rb
@@ -1,0 +1,4 @@
+require 'simple_navbar/render/quick_filter_bar/builder'
+require 'simple_navbar/render/quick_filter_bar/group'
+require 'simple_navbar/render/quick_filter_bar/group_item'
+require 'simple_navbar/render/quick_filter_bar/url_query_string'
