+++ lib/simple_navbar/render/quick_filter_bar/url_query_string.rb
@@ -1,0 +1,15 @@
+module SimpleNavbar
+  module Render
+    module QuickFilterBar
+      class UrlQueryString
+        def self.encode(hash)
+          URI.encode "?#{hash.map { |key,value| "#{key}=#{value}" }*"&"}"
+        end
+
+        def self.decode(query)
+          query.gsub("?","").split("&").map{|str|str.split("=")}.each{|arr|arr[1] = URI.decode arr[1]}.to_h
+        end
+      end
+    end
+  end
+end
