+++ lib/simple_navbar/render/simple_navbar.rb
@@ -1,0 +1,3 @@
+require 'simple_navbar/render/simple_navbar/base/navbar'
+require 'simple_navbar/render/simple_navbar/base/nav_item'
+require 'simple_navbar/render/simple_navbar/default_navbar'
