+++ lib/simple_navbar/render/simple_navtabs.rb
@@ -1,0 +1,2 @@
+require 'simple_navbar/render/simple_navtabs/navtabs'
+require 'simple_navbar/render/simple_navtabs/navtabs_item'
