+++ lib/simple_navbar/render/simple_navtabs/navtabs.rb
@@ -1,0 +1,29 @@
+module SimpleNavbar
+  module Render
+    module SimpleNavtabs
+      class Navtabs
+        def initialize(view, rule)
+          @view = view
+          @rule = rule
+        end
+
+        def navtabs_items
+          @rule.navs.map do |nav|
+            text = nav.options.name
+            url  = nav.options.url
+            controller_items = nav.controller_items
+            ::SimpleNavbar::Render::SimpleNavtabs::NavtabsItem.new(@view, text, url, controller_items)
+          end
+        end
+
+        def render
+          @view.haml_tag :ul, :class => 'nav nav-tabs' do
+            navtabs_items.each do |item|
+              item.render
+            end
+          end
+        end
+      end
+    end
+  end
+end
