--- lib/simple_navbar/simple_navtabs_helpers.rb
@@ -1,14 +1,0 @@
-# -*- encoding : utf-8 -*-
-module SimpleNavbar
-  module SimpleNavtabsHelpers
-    def simple_navtabs(rule)
-      rule    = SimpleNavbar::Rule.get(rule.to_sym)
-      navtabs = SimpleNavbar::Render::Navtabs.new(self, rule)
-
-      capture_haml do
-        navtabs.render
-      end
-    end
-
-  end
-end
